new comp: walls by system family
- improved showing document name on component message
- improved document parameter logic on  DocumentComponent. See note 1

1)
DocumentComponents can decide to define an optional input parameter to get source documents. DocumentComponents now has a helper method to create this document parameter when needed and ensure it is consistent. DocumentComponents inspects the derived components, checks if there is any optional input parameter and if it has valid value. If no optional document parameter is defined or the input is invalid, it provides the active document as the source document to the derived component solver.